# day02-tlias-web-management/Web开发-2024.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Web开发-2024",
    "description": "新版本的Web开发的接口文档",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/emps/{id}": {
      "get": {
        "summary": "根据ID查询",
        "x-apifox-folder": "员工管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "根据ID查询员工",
        "operationId": "getInfoUsingGET_1",
        "tags": [
          "员工管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "example": "1",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-apifox-refs": {},
                  "x-apifox-orders": [
                    "code",
                    "msg",
                    "data"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer",
                      "title": "状态码",
                      "description": "1: 成功 , 0: 失败",
                      "format": "int32",
                      "additionalProperties": false
                    },
                    "msg": {
                      "type": "string",
                      "title": "错误提示信息",
                      "additionalProperties": false,
                      "nullable": true
                    },
                    "data": {
                      "title": "响应数据",
                      "type": "object",
                      "x-apifox-refs": {},
                      "x-apifox-orders": [
                        "id",
                        "username",
                        "password",
                        "name",
                        "gender",
                        "job",
                        "salary",
                        "image",
                        "entryDate",
                        "deptId",
                        "createTime",
                        "updateTime",
                        "exprList"
                      ],
                      "properties": {
                        "id": {
                          "type": "integer",
                          "title": "ID",
                          "format": "int32",
                          "description": "ID",
                          "additionalProperties": false,
                          "nullable": true
                        },
                        "username": {
                          "type": "string",
                          "title": "用户名",
                          "description": "用户名",
                          "additionalProperties": false,
                          "nullable": true
                        },
                        "password": {
                          "type": "string",
                          "title": "密码",
                          "description": "密码",
                          "additionalProperties": false,
                          "nullable": true
                        },
                        "name": {
                          "type": "string",
                          "title": "姓名",
                          "description": "姓名",
                          "additionalProperties": false,
                          "nullable": true
                        },
                        "gender": {
                          "type": "integer",
                          "title": "性别",
                          "description": "性别, 1: 男 , 2: 女",
                          "format": "int32",
                          "additionalProperties": false,
                          "nullable": true
                        },
                        "job": {
                          "type": "integer",
                          "title": "职位",
                          "description": "职位, 1 班主任,2 讲师, 3 学工主管, 4 教研主管, 5 咨询师",
                          "format": "int32",
                          "additionalProperties": false,
                          "nullable": true
                        },
                        "salary": {
                          "type": "integer",
                          "format": "int32",
                          "title": "薪资",
                          "description": "薪资",
                          "additionalProperties": false
                        },
                        "image": {
                          "type": "string",
                          "title": "头像",
                          "description": "头像",
                          "additionalProperties": false,
                          "nullable": true
                        },
                        "entryDate": {
                          "type": "string",
                          "title": "入职日期",
                          "format": "date",
                          "description": "入职日期",
                          "additionalProperties": false,
                          "nullable": true
                        },
                        "deptId": {
                          "type": "integer",
                          "title": "部门ID",
                          "format": "int32",
                          "description": "部门ID",
                          "additionalProperties": false,
                          "nullable": true
                        },
                        "createTime": {
                          "type": "string",
                          "title": "创建时间",
                          "format": "date-time",
                          "description": "创建时间",
                          "additionalProperties": false,
                          "nullable": true
                        },
                        "updateTime": {
                          "type": "string",
                          "title": "修改时间",
                          "format": "date-time",
                          "description": "修改时间",
                          "additionalProperties": false,
                          "nullable": true
                        },
                        "exprList": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/EmpExpr"
                          },
                          "title": "工作经历",
                          "description": "工作经历"
                        }
                      },
                      "required": [
                        "salary"
                      ],
                      "x-apifox-ignore-properties": [],
                      "nullable": true
                    }
                  },
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3161925/apis/api-103410799-run"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmpExpr": {
        "$comment": "JSON Schema for emp_expr table",
        "$id": "emp_expr",
        "title": "",
        "description": "员工工作经历表",
        "type": "object",
        "properties": {
          "id": {
            "$comment": "primary key",
            "description": "ID, 主键",
            "type": "integer",
            "minimum": 1,
            "maximum": 4294967296,
            "title": "ID, 主键"
          },
          "begin": {
            "description": "开始时间",
            "type": "string",
            "format": "date",
            "title": "开始时间"
          },
          "end": {
            "description": "结束时间",
            "type": "string",
            "format": "date",
            "title": "结束时间"
          },
          "company": {
            "description": "公司名字",
            "type": "string",
            "maxLength": 50,
            "title": "公司名字"
          },
          "job": {
            "description": "职位",
            "type": "string",
            "maxLength": 50,
            "title": "职位"
          },
          "empId": {
            "description": "员工ID",
            "type": "integer",
            "minimum": 0,
            "maximum": 4294967296,
            "title": "员工ID"
          }
        },
        "x-apifox-orders": [
          "id",
          "begin",
          "end",
          "company",
          "job",
          "empId"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "Schemas"
      },
...
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "开发环境"
    },
    {
      "url": "http://test-cn.your-api-server.com",
      "description": "测试环境"
    },
    {
      "url": "http://prod-cn.your-api-server.com",
      "description": "正式环境"
    }
  ]
}